Separate report tests into own file
